% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_confidence_HY5_genes}
\alias{high_confidence_HY5_genes}
\title{127 genes from heatmap Figure 3C right: DEGs in HY5ox HY5-VP16 and HY5-SRDX with HY5 binding peak (2000bp upstream to the ATG or in gene body), in all 8 ChIP-seq samples}
\format{
Character vector
}
\usage{
high_confidence_HY5_genes
}
\description{
127 genes from heatmap Figure 3C right: DEGs in HY5ox HY5-VP16 and HY5-SRDX with HY5 binding peak (2000bp upstream to the ATG or in gene body), in all 8 ChIP-seq samples
}
\keyword{datasets}
