% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HY5}
\alias{HY5}
\title{HY5 binding site motif M08407_2.00 from CIS-BP, based on JASPAR}
\format{
Universalmotif object
}
\usage{
HY5
}
\description{
HY5 binding site motif M08407_2.00 from CIS-BP, based on JASPAR
}
\keyword{datasets}
