% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HY5_motif_genes}
\alias{HY5_motif_genes}
\title{HY5 motif genes}
\format{
An object of class \code{character} of length 92.
}
\usage{
HY5_motif_genes
}
\description{
HY5 motif genes
}
\keyword{datasets}
