% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HY5_mutant_comparisons}
\alias{HY5_mutant_comparisons}
\title{log2 FC of gene expression in hy5 mutant and ox lines}
\format{
Data frame
#' \describe{
  \item{Gene}{Arabidopsis AT number}
  \item{OX}{(Overexpression}
  \item{VP16}{Activation domain}
  \item{SRDX}{Transcriptional silencing motif (EAR repressor motif of SUPERMAN gene)}
}
}
\usage{
HY5_mutant_comparisons
}
\description{
log2 FC of gene expression in hy5 mutant and ox lines
}
\keyword{datasets}
