% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HY5_overlap_table.R
\name{HY5_overlap_table}
\alias{HY5_overlap_table}
\title{Output a table of the overlapping of HY5-related genes}
\usage{
HY5_overlap_table(output_file = "HY5_gene_overlap.tsv")
}
\arguments{
\item{output_file}{Output path of the table}
}
\value{
None
}
\description{
Output a table of the overlapping of HY5-related genes
}
