% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HY5_predicted_targets}
\alias{HY5_predicted_targets}
\title{297 HY5_predicted_targets from Burko et al. 2020 Supp Material S2}
\format{
Character vector
}
\usage{
HY5_predicted_targets
}
\description{
297 HY5_predicted_targets from Burko et al. 2020 Supp Material S2
}
\keyword{datasets}
