% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HY5_Venn.R
\name{HY5_Venn}
\alias{HY5_Venn}
\title{Output a Venn diagram of the overlapping of HY5-related genes}
\usage{
HY5_Venn(output_file = "HY5_Venn.pdf")
}
\arguments{
\item{output_file}{Output path of the table}
}
\value{
None
}
\description{
Output a Venn diagram of the overlapping of HY5-related genes
}
