% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenylpropanoid_gene_table}
\alias{phenylpropanoid_gene_table}
\title{Data frame of genes in the phenylpropanoid pathway}
\format{
Data frame
#' \describe{
  \item{AT}{Arabidopsis AT number}
  \item{Direction}{(1:positive; 2:negative; 3:other}
  \item{Mode}{1:enzyme; 2:transcription; 3:protein; 4:other}
  \item{Branch}{1:lignin; 2:flavonoid-anthocyanin; 3:coumarin, 4:cutin-suberin; 5:common}
  \item{Evidence}{1:experiment; 2:prediction; 3:other)}
}
}
\usage{
phenylpropanoid_gene_table
}
\description{
Data frame of genes in the phenylpropanoid pathway
}
\keyword{datasets}
