% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAIR10_Grange}
\alias{TAIR10_Grange}
\title{Grange object from Arabidopsis_thaliana.TAIR10.47.gtf}
\format{
An object of class \code{GRanges} of length 888095.
}
\source{
\url{https://plants.ensembl.org/Arabidopsis_thaliana/Info/Index}
}
\usage{
TAIR10_Grange
}
\description{
Grange object from Arabidopsis_thaliana.TAIR10.47.gtf
}
\keyword{datasets}
